+++ benches/benchmark.rs
@@ -8,9 +8,9 @@ use particular::prelude::*;
 
 struct Dummy;
 
-impl ComputeMethod<Vector, f32> for Dummy {
-    fn compute(&mut self, particles: &[(Vector, f32)]) -> Vec<Vector> {
-        vec![Vector::ZERO; particles.len()]
+impl ComputeMethod<Vec<(glam::Vec3A, f32)>> for Dummy {
+    fn compute(&mut self, storage: Vec<(glam::Vec3A, f32)>) -> Vec<glam::Vec3A> {
+        vec![glam::Vec3A::ZERO; storage.len()]
     }
 }
 
@@ -20,7 +20,10 @@ pub struct Body {
     mu: f32,
 }
 
-fn get_acceleration(bodies: &[Body], cm: &mut impl ComputeMethod<glam::Vec3A, f32>) {
+fn get_acceleration<B>(bodies: &[Body], cm: &mut impl ComputeMethod<B>)
+where
+    B: ParticleStorage<InternalVector = glam::Vec3A, Scalar = f32>,
+{
     let _ = bodies.iter().accelerations(cm).collect::<Vec<_>>();
 }