+++ src/compute_method/mod.rs
@@ -24,18 +24,69 @@ pub mod sequential;
 /// # use glam::Vec3A;
 /// struct AccelerationCalculator;
 ///
-/// impl ComputeMethod<Vec3A, f32> for AccelerationCalculator {
-///     fn compute(&mut self, particles: &[(Vec3A, f32)]) -> Vec<Vec3A> {
+/// impl ComputeMethod<Vec<(Vec3A, f32)>> for AccelerationCalculator {
+///     fn compute(&mut self, particles: Vec<(Vec3A, f32)>) -> Vec<Vec3A> {
 ///     // ...
 /// #       Vec::new()
 ///     }
 /// }
 /// ```
-pub trait ComputeMethod<T, S> {
+pub trait ComputeMethod<B: ParticleStorage> {
     /// Computes the acceleration of the particles.
     ///
     /// The returning vector should contain the acceleration of the particles in the same order they were input.
-    fn compute(&mut self, particles: &[(T, S)]) -> Vec<T>;
+    fn compute(&mut self, storage: B) -> Vec<B::InternalVector>;
+}
+
+/// Storage used by [`ComputeMethod`] to access the particles.
+pub trait ParticleStorage {
+    /// The interal vector used for the stored particles.
+    type InternalVector;
+
+    /// The scalar type used for the stored particles.
+    type Scalar;
+
+    /// Creates an instance of a type implementing [`ParticleStorage`] from a vector of particles.
+    fn new(particles: Vec<(Self::InternalVector, Self::Scalar)>) -> Self;
+}
+
+impl<T, S> ParticleStorage for Vec<(T, S)> {
+    type InternalVector = T;
+
+    type Scalar = S;
+
+    fn new(particles: Vec<(T, S)>) -> Self {
+        particles
+    }
+}
+
+/// Storage for particles with a copy of the massives ones in a second vector.
+pub struct WithMassive<T, S> {
+    /// Particles for which the acceleration is computed.
+    pub particles: Vec<(T, S)>,
+
+    /// Particles used to compute the acceleration of the `particles`.
+    pub massive: Vec<(T, S)>,
+}
+
+impl<T, S> ParticleStorage for WithMassive<T, S>
+where
+    T: Copy,
+    S: Copy + Default + PartialEq,
+{
+    type InternalVector = T;
+
+    type Scalar = S;
+
+    fn new(particles: Vec<(T, S)>) -> Self {
+        let massive: Vec<_> = particles
+            .iter()
+            .filter(|(_, mu)| *mu != S::default())
+            .copied()
+            .collect();
+
+        Self { particles, massive }
+    }
 }
 
 #[cfg(test)]
@@ -43,14 +94,17 @@ pub(crate) mod tests {
     use crate::prelude::*;
     use glam::Vec3A;
 
-    pub fn acceleration_computation<C>(mut cm: C)
+    pub fn acceleration_computation<C, B>(mut cm: C)
     where
-        C: ComputeMethod<Vec3A, f32>,
+        B: ParticleStorage<InternalVector = Vec3A, Scalar = f32>,
+        C: ComputeMethod<B>,
     {
         let massive = vec![(Vec3A::splat(0.0), 2.0), (Vec3A::splat(1.0), 3.0)];
         let massless = vec![(Vec3A::splat(5.0), 0.0)];
 
-        let computed = cm.compute(&[massive.clone(), massless.clone()].concat()[..]);
+        let computed = cm.compute(ParticleStorage::new(
+            [massive.clone(), massless.clone()].concat(),
+        ));
 
         for (&point_mass1, computed) in massive.iter().chain(massless.iter()).zip(computed) {
             let mut acceleration = Vec3A::ZERO;